moved initialization into file scope to ensure it's ready before react is
