+++ libs/shared/src/components/UX/Numeric/styles.module.less
@@ -39,5 +39,6 @@
   }
   &.years span::after {
     content: 'yrs';
+    word-break: normal;
   }
 }
